Updated to .net standard 2.0 Updated RavenDB to 3.x
