+++ src/CR.ViewModels.sln
@@ -1,0 +1,59 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.27428.2037
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Core", "Core\Core.csproj", "{02B3D2E1-0541-46E7-87EA-8A49F5755969}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Persistence", "Persistence", "{0B32DCBE-5335-4D6E-9385-D069BBBA8634}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Persistence.ApplicationState", "Persistence.ApplicationState\Persistence.ApplicationState.csproj", "{A4A3A5EF-3F85-47A5-A39C-DF936CAE60AC}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Persistence.Memory", "Persistence.Memory\Persistence.Memory.csproj", "{33902919-1EAA-4673-85FD-102267E2729D}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Persistence.RavenDB", "Persistence.RavenDB\Persistence.RavenDB.csproj", "{B28BB5C0-B434-438C-8770-BBFF710A2708}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{9B7A4714-4B23-4675-82F9-F85C6E23E01E}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tests", "Tests\Tests.csproj", "{56CE25A6-B966-4880-A2A9-7399B40235D6}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{02B3D2E1-0541-46E7-87EA-8A49F5755969}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{02B3D2E1-0541-46E7-87EA-8A49F5755969}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{02B3D2E1-0541-46E7-87EA-8A49F5755969}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{02B3D2E1-0541-46E7-87EA-8A49F5755969}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A4A3A5EF-3F85-47A5-A39C-DF936CAE60AC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A4A3A5EF-3F85-47A5-A39C-DF936CAE60AC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A4A3A5EF-3F85-47A5-A39C-DF936CAE60AC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A4A3A5EF-3F85-47A5-A39C-DF936CAE60AC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{33902919-1EAA-4673-85FD-102267E2729D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{33902919-1EAA-4673-85FD-102267E2729D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{33902919-1EAA-4673-85FD-102267E2729D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{33902919-1EAA-4673-85FD-102267E2729D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B28BB5C0-B434-438C-8770-BBFF710A2708}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B28BB5C0-B434-438C-8770-BBFF710A2708}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B28BB5C0-B434-438C-8770-BBFF710A2708}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B28BB5C0-B434-438C-8770-BBFF710A2708}.Release|Any CPU.Build.0 = Release|Any CPU
+		{56CE25A6-B966-4880-A2A9-7399B40235D6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{56CE25A6-B966-4880-A2A9-7399B40235D6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{56CE25A6-B966-4880-A2A9-7399B40235D6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{56CE25A6-B966-4880-A2A9-7399B40235D6}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{A4A3A5EF-3F85-47A5-A39C-DF936CAE60AC} = {0B32DCBE-5335-4D6E-9385-D069BBBA8634}
+		{33902919-1EAA-4673-85FD-102267E2729D} = {0B32DCBE-5335-4D6E-9385-D069BBBA8634}
+		{B28BB5C0-B434-438C-8770-BBFF710A2708} = {0B32DCBE-5335-4D6E-9385-D069BBBA8634}
+		{56CE25A6-B966-4880-A2A9-7399B40235D6} = {9B7A4714-4B23-4675-82F9-F85C6E23E01E}
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {FD9674CB-AD88-4076-B2ED-1EA637F09BC8}
+	EndGlobalSection
+EndGlobal
