+++ src/Persistence.RavenDB/RavenDBViewModelReader.cs
@@ -1,12 +1,9 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Linq;
-using System.Linq.Expressions;
 using CR.ViewModels.Core;
-using CR.ViewModels.Core.Exceptions;
 using Raven.Client;
 
-namespace CR.ViewModels.Persistance.RavenDB
+namespace CR.ViewModels.Persistence.RavenDB
 {
     public class RavenDBViewModelReader : IViewModelReader
     {
