+++ src/core/Exceptions/EntityNotFoundException.cs
@@ -1,9 +1,5 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Runtime.Serialization;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace CR.ViewModels.Core.Exceptions
 {
