--- src/core/core.nuspec
@@ -1,17 +1,0 @@
-<?xml version="1.0"?>
-<package >
-  <metadata>
-    <id>$id$</id>
-    <version>$version$</version>
-    <title>$title$</title>
-    <authors>Rich Knight,Jon Bradford</authors>
-    <owners>Rich Knight,Jon Bradford</owners>
-    <licenseUrl>https://raw.githubusercontent.com/cognisant/cr-viewmodels/master/LICENSE</licenseUrl>
-    <projectUrl>https://github.com/cognisant/cr-viewmodels</projectUrl>
-    <iconUrl>https://raw.githubusercontent.com/cognisant/cr-viewmodels/master/logo.png</iconUrl>
-    <requireLicenseAcceptance>false</requireLicenseAcceptance>
-    <description>A set of .net interfaces for reading/writing CQRS ViewModels in a generic way</description>
-    <releaseNotes>Initial Version</releaseNotes>
-    <copyright>Copyright 2015</copyright>
-  </metadata>
-</package>
