--- src/cr-viewmodels.sln
@@ -1,53 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2013
-VisualStudioVersion = 12.0.30501.0
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "core", "core\core.csproj", "{8BFEA4BE-C00F-4C54-9AD2-6BCD70C4A0D9}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "persistance.memory", "persistance.memory\persistance.memory.csproj", "{D5DCB132-A8C5-4057-AA2E-3E12EFD63CD1}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "tests", "tests\tests.csproj", "{1E8BB448-FD81-4CB0-A67A-D2EC5D8C98AC}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "persistance.applicationstate", "persistance.applicationstate\persistance.applicationstate.csproj", "{186E570A-1A2B-4064-B711-4745B3389251}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".nuget", ".nuget", "{348EF9EB-A6A2-4CA8-856E-095BCAD021FB}"
-	ProjectSection(SolutionItems) = preProject
-		.nuget\NuGet.Config = .nuget\NuGet.Config
-		.nuget\NuGet.exe = .nuget\NuGet.exe
-		.nuget\NuGet.targets = .nuget\NuGet.targets
-	EndProjectSection
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "persistance.ravendb", "persistance.ravendb\persistance.ravendb.csproj", "{898F2107-34AD-459B-9B33-A722F3DF9E40}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{8BFEA4BE-C00F-4C54-9AD2-6BCD70C4A0D9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8BFEA4BE-C00F-4C54-9AD2-6BCD70C4A0D9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8BFEA4BE-C00F-4C54-9AD2-6BCD70C4A0D9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8BFEA4BE-C00F-4C54-9AD2-6BCD70C4A0D9}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D5DCB132-A8C5-4057-AA2E-3E12EFD63CD1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D5DCB132-A8C5-4057-AA2E-3E12EFD63CD1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D5DCB132-A8C5-4057-AA2E-3E12EFD63CD1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D5DCB132-A8C5-4057-AA2E-3E12EFD63CD1}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1E8BB448-FD81-4CB0-A67A-D2EC5D8C98AC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1E8BB448-FD81-4CB0-A67A-D2EC5D8C98AC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1E8BB448-FD81-4CB0-A67A-D2EC5D8C98AC}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1E8BB448-FD81-4CB0-A67A-D2EC5D8C98AC}.Release|Any CPU.Build.0 = Release|Any CPU
-		{186E570A-1A2B-4064-B711-4745B3389251}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{186E570A-1A2B-4064-B711-4745B3389251}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{186E570A-1A2B-4064-B711-4745B3389251}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{186E570A-1A2B-4064-B711-4745B3389251}.Release|Any CPU.Build.0 = Release|Any CPU
-		{898F2107-34AD-459B-9B33-A722F3DF9E40}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{898F2107-34AD-459B-9B33-A722F3DF9E40}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{898F2107-34AD-459B-9B33-A722F3DF9E40}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{898F2107-34AD-459B-9B33-A722F3DF9E40}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
