--- src/persistance.memory/persistance.memory.nuspec
@@ -1,20 +1,0 @@
-﻿<?xml version="1.0"?>
-<package >
-  <metadata>
-    <id>$id$</id>
-    <version>$version$</version>
-    <title>$title$</title>
-    <authors>Rich Knight,Jon Bradford</authors>
-    <owners>Rich Knight,Jon Bradford</owners>
-    <licenseUrl>https://raw.githubusercontent.com/cognisant/cr-viewmodels/master/LICENSE</licenseUrl>
-    <projectUrl>https://github.com/cognisant/cr-viewmodels</projectUrl>
-    <iconUrl>https://raw.githubusercontent.com/cognisant/cr-viewmodels/master/logo.png</iconUrl>
-    <requireLicenseAcceptance>false</requireLicenseAcceptance>
-    <description>In memory persistance for CR ViewModels</description>
-    <releaseNotes>Initial Version</releaseNotes>
-    <copyright>Copyright 2015</copyright>
-    <dependencies>
-      <dependency id="CR.ViewModels.Core" version="1.0.0" />
-    </dependencies>
-  </metadata>
-</package>
