--- src/persistance.ravendb/packages.config
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Newtonsoft.Json" version="4.0.8" targetFramework="net45" />
-  <package id="NLog" version="2.0.0.2000" targetFramework="net45" />
-  <package id="RavenDB.Client" version="1.0.992" targetFramework="net45" />
-</packages>
