+++ src/tests/packages.config
@@ -1,9 +1,13 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Newtonsoft.Json" version="4.0.8" targetFramework="net45" />
-  <package id="NLog" version="2.0.0.2000" targetFramework="net45" />
-  <package id="NUnit" version="2.6.2" targetFramework="net45" />
-  <package id="RavenDB.Client" version="1.0.992" targetFramework="net45" />
-  <package id="RavenDB.Database" version="1.0.992" targetFramework="net45" />
-  <package id="RavenDB.Embedded" version="1.0.992" targetFramework="net45" />
+  <package id="Microsoft.CSharp" version="4.4.1" targetFramework="net471" />
+  <package id="Newtonsoft.Json" version="11.0.2" targetFramework="net471" />
+  <package id="NUnit" version="3.10.1" targetFramework="net471" />
+  <package id="RavenDB.Client" version="3.5.5" targetFramework="net471" />
+  <package id="RavenDB.Database" version="3.5.5" targetFramework="net471" />
+  <package id="System.Buffers" version="4.4.0" targetFramework="net471" />
+  <package id="System.Runtime.CompilerServices.Unsafe" version="4.4.0" targetFramework="net471" />
+  <package id="System.Runtime.InteropServices.RuntimeInformation" version="4.3.0" targetFramework="net471" />
+  <package id="System.Security.Cryptography.Cng" version="4.4.0" targetFramework="net471" />
+  <package id="System.Threading.Tasks.Extensions" version="4.4.0" targetFramework="net471" />
 </packages>
